+++ OpenNos.Data/LogDropDTO.cs
@@ -20,9 +20,9 @@ namespace OpenNos.Data
 
         public short Map { get; set; }
 
-        public byte X { get; set; }
+        public short X { get; set; }
 
-        public byte Y { get; set; }
+        public short Y { get; set; }
 
         public string IpAddress { get; set; }
